refactor ziogrpc
