+++ ziogrpc-plugin-common/src/main/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/common/ZioGrpcWitnessConstant.scala
@@ -1,7 +1,4 @@
-package org.bitlap.skywalking.apm.plugin.ziogrpc.v06rcx
-
-import net.bytebuddy.matcher.ElementMatchers
-import net.bytebuddy.matcher.ElementMatchers.takesArguments
+package org.bitlap.skywalking.apm.plugin.ziogrpc.common
 
 /** @author
  *    梦境迷离
