+++ ziogrpc-v06rcx-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/v06rcx/interceptor/ZioGrpcClientInterceptor.scala
@@ -12,7 +12,9 @@ import zio.*
 import org.apache.skywalking.apm.agent.core.context.ContextManager
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.*
 import org.bitlap.skywalking.apm.plugin.common.Utils
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06rcx.forward.TracingClientCall
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.*
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.Constants.*
+import org.bitlap.skywalking.apm.plugin.ziogrpc.v06rcx.TracingClientCall
 
 /** @author
  *    梦境迷离
