+++ ziogrpc-v06rcx-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/v06rcx/interceptor/ZioGrpcServerCallConstructorInterceptor.scala
@@ -4,7 +4,8 @@ import io.grpc.*
 
 import org.apache.skywalking.apm.agent.core.context.ContextManager
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.*
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06rcx.OperationContext
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.*
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.Constants.*
 
 /** @author
  *    梦境迷离
