+++ ziogrpc-v06testx-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/v06testx/interceptor/ZioGrpcClientInterceptor.scala
@@ -12,7 +12,7 @@ import zio.*
 import org.apache.skywalking.apm.agent.core.context.ContextManager
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.*
 import org.bitlap.skywalking.apm.plugin.common.Utils
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.forward.TracingClientCall
+import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.TracingClientCall
 
 /** @author
  *    梦境迷离
