+++ ziogrpc-v06testx-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/v06testx/interceptor/ZioGrpcServerInterceptor.scala
@@ -18,9 +18,9 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.*
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine
 import org.apache.skywalking.apm.util.StringUtil
 import org.bitlap.skywalking.apm.plugin.common.*
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.*
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.Constants.ZIO_GRPC
-import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.forward.TracingServerCallListener
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.*
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.Constants.*
+import org.bitlap.skywalking.apm.plugin.ziogrpc.common.listener.*
 
 /** @author
  *    梦境迷离
