+++ ziogrpc-v06testx-plugin/src/test/scala/org/bitlap/skywalking/apm/plugin/ziogrpc/v06testx/ZioGrpcServerCallInstrumentationSpec.scala
@@ -10,6 +10,7 @@ import io.grpc.*
 
 import zio.Trace
 
+import org.bitlap.skywalking.apm.plugin.ziogrpc.v06testx.define.*
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
